Safari-609.1.11.1/LayoutTests/http/tests/webfont/fallback-font-while-loading-expected.txt

This test checks that the fallback font is used for layout while a webfont is loading.

Target:
A text to be measured.
Reference:
A text to be measured.
PASS: The width of target text and reference text is the same.
